+++ src/main/java/frc/robot/pearce/RobotContainer.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot;
+package frc.robot.pearce;
 
 import com.pathplanner.lib.auto.AutoBuilder;
 import com.pathplanner.lib.auto.NamedCommands;
@@ -16,14 +16,15 @@ import edu.wpi.first.wpilibj.smartdashboard.SendableChooser;
 import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.Commands;
+import frc.robot.Constants;
 import frc.robot.Constants.HIDConstants;
 import frc.robot.common.annotations.Robot;
 import frc.robot.common.components.RobotUtils;
 import frc.robot.common.interfaces.IRobotContainer;
-import frc.robot.common.subsystems.CBSSubsystem;
-import frc.robot.common.subsystems.DeepClimbSubsystem;
-import frc.robot.common.subsystems.ElevatorSubsystem;
-import frc.robot.common.subsystems.ScoringSubsystem;
+import frc.robot.pearce.subsystems.CBSSubsystem;
+import frc.robot.pearce.subsystems.DeepClimbSubsystem;
+import frc.robot.pearce.subsystems.ElevatorSubsystem;
+import frc.robot.pearce.subsystems.ScoringSubsystem;
 import frc.robot.common.subsystems.drive.SwerveDriveSubsystem;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
