+++ src/main/java/frc/robot/pearce/subsystems/CBSSubsystem.java
@@ -1,4 +1,4 @@
-package frc.robot.common.subsystems;
+package frc.robot.pearce.subsystems;
 
 import com.revrobotics.RelativeEncoder;
 import com.revrobotics.spark.SparkBase;
