+++ src/main/java/frc/robot/pearce/subsystems/DeepClimbSubsystem.java
@@ -1,4 +1,4 @@
-package frc.robot.common.subsystems;
+package frc.robot.pearce.subsystems;
 
 import com.revrobotics.spark.SparkFlex;
 import com.revrobotics.spark.SparkLowLevel;
