+++ src/main/java/frc/robot/pearce/subsystems/ScoringSubsystem.java
@@ -1,4 +1,4 @@
-package frc.robot.common.subsystems;
+package frc.robot.pearce.subsystems;
 
 import com.revrobotics.spark.SparkBase;
 import com.revrobotics.spark.SparkFlex;
